moved team to blueprints
