+++ api/api/routes/team.py
@@ -1,0 +1,25 @@
+from flask import Flask, request, session, send_from_directory, render_template
+from flask import Blueprint
+import api
+
+from api.common import WebSuccess, WebError
+from api.annotations import api_wrapper, require_login, require_teacher, require_admin, check_csrf
+from api.annotations import block_before_competition, block_after_competition
+from api.annotations import log_action
+
+blueprint = Blueprint("team_api", __name__, template_folder="templates")
+
+@blueprint.route('', methods=['GET'])
+@api_wrapper
+@require_login
+def team_information_hook():
+    return WebSuccess(data=api.team.get_team_information())
+
+@blueprint.route('/score', methods=['GET'])
+@api_wrapper
+@require_login
+def get_team_score_hook():
+    score = api.stats.get_score(tid=api.user.get_user()['tid'])
+    if score is not None:
+        return WebSuccess(data={'score': score})
+    return WebError("There was an error retrieving your score.")
